manual merge homechat
